requirements.txt
asgiref==3.7.2
Django==4.2.10
dj-database-url==2.1.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-import-export==3.3.7
djangorestframework==3.14.0
pillow==10.2.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1